+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraKeyValueServiceImpl.java
@@ -1730,18 +1730,20 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
     }
 
     @Override
-    public void deleteAllTimestamps(TableReference tableRef, Map<Cell, Long> maxTimestampExclusiveByCell) {
+    public void deleteAllTimestamps(TableReference tableRef, Map<Cell, Long> maxTimestampExclusiveByCell,
+            boolean deleteSentinels) {
         Map<InetSocketAddress, Map<Cell, Long>> keysByHost = HostPartitioner.partitionMapByHost(
                 clientPool, maxTimestampExclusiveByCell.entrySet());
         for (Map.Entry<InetSocketAddress, Map<Cell, Long>> entry : keysByHost.entrySet()) {
-            deleteAllTimestampsOnSingleHost(tableRef, entry.getKey(), entry.getValue());
+            deleteAllTimestampsOnSingleHost(tableRef, entry.getKey(), entry.getValue(), deleteSentinels);
         }
     }
 
     public void deleteAllTimestampsOnSingleHost(
             TableReference tableRef,
             InetSocketAddress host,
-            Map<Cell, Long> maxTimestampExclusiveByCell) {
+            Map<Cell, Long> maxTimestampExclusiveByCell,
+            boolean deleteSentinels) {
         if (maxTimestampExclusiveByCell.isEmpty()) {
             return;
         }
@@ -1751,7 +1753,7 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
 
                 @Override
                 public Void apply(CassandraClient client) throws Exception {
-                    insertRangeTombstones(client, maxTimestampExclusiveByCell, tableRef);
+                    insertRangeTombstones(client, maxTimestampExclusiveByCell, tableRef, deleteSentinels);
                     return null;
                 }
 
@@ -1770,19 +1772,24 @@ public class CassandraKeyValueServiceImpl extends AbstractKeyValueService implem
     }
 
     private void insertRangeTombstones(CassandraClient client, Map<Cell, Long> maxTimestampExclusiveByCell,
-            TableReference tableRef) throws TException {
+            TableReference tableRef, boolean deleteSentinel) throws TException {
         MutationMap mutationMap = new MutationMap();
 
         maxTimestampExclusiveByCell.forEach((cell, maxTimestampExclusive) -> {
-            Mutation mutation = Mutations.rangeTombstoneForColumn(
-                    cell.getColumnName(),
-                    maxTimestampExclusive);
+            Mutation mutation = getMutation(cell, maxTimestampExclusive, deleteSentinel);
 
             mutationMap.addMutationForCell(cell, tableRef, mutation);
         });
 
         wrappingQueryRunner.batchMutate("deleteAllTimestamps", client, ImmutableSet.of(tableRef), mutationMap,
                 deleteConsistency);
+    }
+
+    private Mutation getMutation(Cell cell, long maxTimestampExclusive, boolean deleteSentinel) {
+        if (deleteSentinel) {
+            return Mutations.rangeTombstoneIncludingSentinelForColumn(cell.getColumnName(), maxTimestampExclusive);
+        }
+        return Mutations.rangeTombstoneForColumn(cell.getColumnName(), maxTimestampExclusive);
     }
 
     /**